+++ Vulkan/include/Queue.h
@@ -38,12 +38,48 @@ namespace vulkan{
 			EVO_NODISCARD auto native() const -> VkQueue { return this->queue; }
 
 
-			auto submit(evo::ArrayProxy<VkSubmitInfo> submit_infos, VkFence fence) -> evo::Result<> {
-				const VkResult result = 
-					vkQueueSubmit(this->queue, uint32_t(submit_infos.size()), submit_infos.data(), fence);
-				if(utils::checkResult(result, "vkQueueSubmit").isError()){ return evo::resultError; }
-				return evo::Result<>();
+
+			auto createSubmitInfo(
+				evo::ArrayProxy<VkSemaphore> wait_semaphore,
+				evo::ArrayProxy<VkPipelineStageFlags> pipeline_stage_mask,
+				evo::ArrayProxy<class CommandBuffer>
+			) -> VkSubmitInfo {
+				evo::debugAssert(
+					wait_semaphore.size() == pipeline_stage_mask.size(),
+					"Num of pipeline_stage_mask must be the same as wait_semaphore"
+				);
 			}
+
+
+
+			struct SubmitInfo{
+				evo::ArrayProxy<VkSemaphore> wait_semaphores;
+				evo::ArrayProxy<VkPipelineStageFlags> pipeline_stage_masks;
+				evo::ArrayProxy<class CommandBuffer> command_buffers;
+				evo::ArrayProxy<VkSemaphore> signal_semaphores;
+
+				SubmitInfo(
+					evo::ArrayProxy<VkSemaphore> _wait_semaphores,
+					evo::ArrayProxy<VkPipelineStageFlags> _pipeline_stage_masks,
+					evo::ArrayProxy<class CommandBuffer> _command_buffers,
+					evo::ArrayProxy<VkSemaphore> _signal_semaphores
+				) : 
+					wait_semaphores(_wait_semaphores),
+					pipeline_stage_masks(_pipeline_stage_masks),
+					command_buffers(_command_buffers),
+					signal_semaphores(_signal_semaphores)
+				{
+					evo::debugAssert(
+						this->wait_semaphores.empty() 
+							|| this->wait_semaphores.size() == this->pipeline_stage_masks.size(),
+						"if pipeline_stage_masks not empty, "
+							"num of pipeline_stage_masks must be the same as wait_semaphores"
+					);
+				}
+			};
+
+
+			auto submit(evo::ArrayProxy<SubmitInfo> submit_infos, class Fence& fence) -> evo::Result<>;
 
 			auto waitIdle() -> void { vkQueueWaitIdle(this->queue); }